plugin testing framework works!
