+++ gulp/content.js
@@ -26,3 +26,4 @@ function contentTask(cwd) {
 
 gulp.task("content-sdk", function() { return contentTask("sdk"); });
 gulp.task("content-cli", function() { return contentTask("cli"); });
+gulp.task("content-plugin", function() { return contentTask("plugin-testing-framework"); });
