+++ package.json
@@ -26,7 +26,7 @@
     "supertest": "^1.0.1",
     "through2": "^2.0.0",
     "ts-jsx-loader": "^0.2.1",
-    "typescript": "1.6.0",
+    "typescript": "1.8.10",
     "webpack": "^1.10.1",
     "which": "^1.1.1"
   }
