+++ plugin-testing-framework/.npmignore
@@ -1,3 +1,4 @@
 .npmignore
 .gitignore
 node_modules/*
+definitions/*
