+++ plugin-testing-framework/definitions/replace.d.ts
