+++ plugin-testing-framework/script/index.ts
@@ -1,0 +1,6 @@
+import * as Platform from "./platform";
+import * as PluginTestingFramework from "./test";
+import { ProjectManager } from "./projectManager";
+import * as ServerUtil from "./serverUtil";
+import * as TestUtil from "./testUtil";
+export { Platform, PluginTestingFramework, ProjectManager, ServerUtil, TestUtil };
