+++ plugin-testing-framework/script/platform.ts
@@ -1,9 +1,6 @@
-/// <reference path="../typings/node.d.ts" />
-
 "use strict";
 
 import path = require("path");
-import ProjectManager = require("./projectManager");
 import tu = require("./testUtil");
 import Q = require("q");
 
@@ -361,14 +358,14 @@ export class AndroidEmulatorManager implements IEmulatorManager {
      * Launches an already installed application by app id.
      */
     launchInstalledApplication(appId: string): Q.Promise<string> {
-        return ProjectManager.ProjectManager.execChildProcess("adb shell monkey -p " + appId + " -c android.intent.category.LAUNCHER 1");
+        return tu.TestUtil.getProcessOutput("adb shell monkey -p " + appId + " -c android.intent.category.LAUNCHER 1");
     }
     
     /**
      * Ends a running application given its app id.
      */
     endRunningApplication(appId: string): Q.Promise<string> {
-        return ProjectManager.ProjectManager.execChildProcess("adb shell am force-stop " + appId);
+        return tu.TestUtil.getProcessOutput("adb shell am force-stop " + appId);
     }
     
     /**
@@ -406,14 +403,14 @@ export class AndroidEmulatorManager implements IEmulatorManager {
      */
     prepareEmulatorForTest(appId: string): Q.Promise<string> {
         return this.endRunningApplication(appId)
-            .then(() => { return ProjectManager.ProjectManager.execChildProcess("adb shell pm clear " + appId); });
+            .then(() => { return tu.TestUtil.getProcessOutput("adb shell pm clear " + appId); });
     }
     
     /**
      * Uninstalls the app from the emulator.
      */
     uninstallApplication(appId: string): Q.Promise<string> {
-        return ProjectManager.ProjectManager.execChildProcess("adb uninstall " + appId);
+        return tu.TestUtil.getProcessOutput("adb uninstall " + appId);
     }
 }
